chore: sync-data-handler tests correctly name the test file
